Eliminate deprecated numpy type aliases.

+++ nion/eels_analysis/test/EELSAnalysisFunctions_test.py
@@ -89,7 +89,7 @@ class TestEELSAnalysisFunctions(unittest.TestCase):
 
     def test_signal_and_background_shape_are_consistent_1d(self):
         calibration = Calibration.Calibration(418.92, 0.97, 'eV')
-        data_and_metadata = DataAndMetadata.DataAndMetadata.from_data(numpy.ones((2048, ), numpy.float), dimensional_calibrations=[calibration])
+        data_and_metadata = DataAndMetadata.DataAndMetadata.from_data(numpy.ones((2048, ), float), dimensional_calibrations=[calibration])
         fit_range = 0.2, 0.3
         signal_range = 0.4, 0.5
         signal = eels_analysis.extract_original_signal(data_and_metadata, [fit_range], signal_range)
@@ -111,7 +111,7 @@ class TestEELSAnalysisFunctions(unittest.TestCase):
     def test_background_signal_has_correct_calibration_offset(self):
         calibration = Calibration.Calibration(200.0, 2.0, 'eV')
         spectrum_length = 1000
-        data_and_metadata = DataAndMetadata.DataAndMetadata.from_data(numpy.ones((spectrum_length,), numpy.float), dimensional_calibrations=[calibration])
+        data_and_metadata = DataAndMetadata.DataAndMetadata.from_data(numpy.ones((spectrum_length,), float), dimensional_calibrations=[calibration])
         background = eels_analysis.calculate_background_signal(data_and_metadata, [(0.2, 0.3)], (0.4, 0.5))
         self.assertEqual(data_and_metadata.dimensional_calibrations[0], calibration)  # dummy check
         self.assertAlmostEqual(background.dimensional_calibrations[0].offset, 0.2 * spectrum_length * calibration.scale + calibration.offset)
@@ -121,7 +121,7 @@ class TestEELSAnalysisFunctions(unittest.TestCase):
     def test_make_signal_like_puts_data_in_correct_place(self):
         calibration = Calibration.Calibration(200.0, 2.0, 'eV')
         spectrum_length = 1000
-        data_and_metadata = DataAndMetadata.DataAndMetadata.from_data(numpy.ones((spectrum_length,), numpy.float), dimensional_calibrations=[calibration])
+        data_and_metadata = DataAndMetadata.DataAndMetadata.from_data(numpy.ones((spectrum_length,), float), dimensional_calibrations=[calibration])
         signal = eels_analysis.extract_original_signal(data_and_metadata, [(0.2, 0.3)], (0.4, 0.5))
         signal = DataAndMetadata.DataAndMetadata.from_data(numpy.ones(300, ), signal.intensity_calibration, signal.dimensional_calibrations)
         expanded = eels_analysis.make_signal_like(signal, data_and_metadata)
@@ -137,7 +137,7 @@ class TestEELSAnalysisFunctions(unittest.TestCase):
         spectrum_length = 1000
         w, h = 20, 20
         electron_shell = PeriodicTable.ElectronShell(1, 1, 0)
-        data_and_metadata = DataAndMetadata.DataAndMetadata.from_data(numpy.ones((spectrum_length, w, h), numpy.float), dimensional_calibrations=[calibration_y, calibration_x, calibration])
+        data_and_metadata = DataAndMetadata.DataAndMetadata.from_data(numpy.ones((spectrum_length, w, h), float), dimensional_calibrations=[calibration_y, calibration_x, calibration])
         mapped = eels_analysis.map_background_subtracted_signal(data_and_metadata, electron_shell, [(0.2, 0.3)], (0.4, 0.5))
         self.assertEqual(len(mapped.dimensional_shape), 2)
         self.assertEqual(len(mapped.dimensional_calibrations), 2)
